+++ gen/generator.go
@@ -106,6 +106,7 @@ type YamlConfig struct {
 	NoDeleteAttributes      bool                  `yaml:"no_delete_attributes"`
 	DefaultDeleteAttributes bool                  `yaml:"default_delete_attributes"`
 	Wait                    bool                  `yaml:"wait"`
+	RequestTimeout          int64                 `yaml:"request_timeout"`
 	TestTags                []string              `yaml:"test_tags"`
 	SkipMinimumTest         bool                  `yaml:"skip_minimum_test"`
 	NoAugmentConfig         bool                  `yaml:"no_augment_config"`